Serve on all interfaces by default
This commit changes the default value of the `--ip` option to use
Node's internal resolution. It means that it will try to bind to
the unspecified IPv6 address `::` if available or else fall back to the
unspecified IPv4 address `0.0.0.0`. In most systems, `::` also binds
to `0.0.0.0` which effectively allows to listen on all interfaces.

Closes hexojs/hexo-server#35