Fix default server config to listen on both IPv4 and IPv6
This commit is a fix following the PR #39. This PR changed the
default behavior of the server to listen on all interfaces (both
IPv4 and IPv6). This commit was missing the update of the default
configuration, causing the server to still listen only to IPv4 by
default. The current commit fixes this error by using `undefined`
(Node's default) for the IP.

- See hexojs/hexo-server#35
- See hexojs/hexo-server#39